make the callback parameter optional
