+++ lso/routes/execute.py
@@ -5,11 +5,10 @@ from pathlib import Path
 from typing import Annotated
 
 from fastapi import APIRouter, HTTPException, status
-from pydantic import AfterValidator, BaseModel, HttpUrl, model_validator
+from pydantic import AfterValidator, BaseModel, HttpUrl
 
-from lso.execute import get_executable_path, run_executable_async
+from lso.execute import get_executable_path, run_executable_async, run_executable_sync
 from lso.schema import ExecutableRunResponse
-from lso.utils import run_executable_sync
 
 router = APIRouter()
 
@@ -40,21 +39,12 @@ class ExecutableRunParams(BaseModel):
     callback: HttpUrl | None = None
     is_async: bool = True
 
-    @model_validator(mode="after")
-    def ensure_callback_for_async(self) -> "ExecutableRunParams":
-        """Ensure that a callback URL is provided if the task is asynchronous."""
-        if self.is_async and self.callback is None:
-            msg = "`callback` is required when `is_async` is true"
-            raise ValueError(msg)
-
-        return self
-
 
 @router.post("/", response_model=ExecutableRunResponse, status_code=status.HTTP_201_CREATED)
 async def run_executable_endpoint(params: ExecutableRunParams) -> ExecutableRunResponse:
     """Dispatch a task to run an arbitrary executable."""
     if params.is_async:
-        job_id = run_executable_async(params.executable_name, params.args, params.callback)  # type: ignore[arg-type]
+        job_id = run_executable_async(params.executable_name, params.args, params.callback)
         return ExecutableRunResponse(job_id=job_id)
 
     job_id = uuid.uuid4()
