+++ lso/tasks.py
@@ -27,7 +27,6 @@ from starlette import status
 
 from lso.config import settings
 from lso.schema import ExecutableRunResponse
-from lso.utils import run_executable_sync
 from lso.worker import RUN_EXECUTABLE, RUN_PLAYBOOK, celery
 
 logger = logging.getLogger(__name__)
@@ -68,31 +67,34 @@ def run_playbook_proc_task(
 
 
 @celery.task(name=RUN_EXECUTABLE)  # type: ignore[misc]
-def run_executable_proc_task(job_id: str, executable_path: str, args: list[str], callback: str) -> None:
+def run_executable_proc_task(job_id: str, executable_path: str, args: list[str], callback: str | None) -> None:
     """Celery task to run an arbitrary executable and notify via callback.
 
-    Executes the executable with the provided arguments and posts back the output and status.
+    Executes the executable with the provided arguments and posts back the result if a callback URL is provided.
     """
+    from lso.execute import run_executable_sync  # noqa: PLC0415
+
     msg = f"Executing executable: {executable_path} with args: {args}, callback: {callback}"
     logger.info(msg)
     result = run_executable_sync(executable_path, args)
 
-    payload = ExecutableRunResponse(
-        job_id=UUID(job_id),
-        result=result,
-    ).model_dump(mode="json")
+    if callback:
+        payload = ExecutableRunResponse(
+            job_id=UUID(job_id),
+            result=result,
+        ).model_dump(mode="json")
 
-    def _raise_callback_error(message: str, error: Exception | None = None) -> None:
-        if error:
-            raise CallbackFailedError(message) from error
-        raise CallbackFailedError(message)
+        def _raise_callback_error(message: str, error: Exception | None = None) -> None:
+            if error:
+                raise CallbackFailedError(message) from error
+            raise CallbackFailedError(message)
 
-    try:
-        response = requests.post(str(callback), json=payload, timeout=settings.REQUEST_TIMEOUT_SEC)
-        if not (status.HTTP_200_OK <= response.status_code < status.HTTP_300_MULTIPLE_CHOICES):
-            msg = f"Callback failed: {response.text}, url: {callback}"
-            _raise_callback_error(msg)
-    except Exception as e:
-        error_msg = f"Callback error: {e}"
-        logger.exception(error_msg)
-        _raise_callback_error(error_msg, e)
+        try:
+            response = requests.post(str(callback), json=payload, timeout=settings.REQUEST_TIMEOUT_SEC)
+            if not (status.HTTP_200_OK <= response.status_code < status.HTTP_300_MULTIPLE_CHOICES):
+                msg = f"Callback failed: {response.text}, url: {callback}"
+                _raise_callback_error(msg)
+        except Exception as e:
+            error_msg = f"Callback error: {e}"
+            logger.exception(error_msg)
+            _raise_callback_error(error_msg, e)
